pyproject.toml
[project]
name = "pozik-bot"
version = "0.1.0"
description = "Telegram userbot for interacting with multiple AI models."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "docx>=0.2.4",
    "googlesearch-python>=1.3.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "telethon>=1.42.0",
]


[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "flake8-pyproject>=1.2.4",
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "wemake-python-styleguide>=1.5.0",
]

[tool.mypy]
strict = true
follow_imports = "silent"

[tool.ruff]
# Ruff config: https://docs.astral.sh/ruff/settings
target-version = "py313"
line-length = 100

preview = true
fix = true
format.quote-style = "single"
format.docstring-code-format = false
lint.select = [
  "A",    # flake8-builtins
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "C90",  # maccabe
  "COM",  # flake8-commas
  "D",    # pydocstyle
  "DTZ",  # flake8-datetimez
  "E",    # pycodestyle
  "ERA",  # flake8-eradicate
  "EXE",  # flake8-executable
  "F",    # pyflakes
  "FBT",  # flake8-boolean-trap
  "FLY",  # pyflint
  "FURB", # refurb
  "G",    # flake8-logging-format
  "I",    # isort
  "ICN",  # flake8-import-conventions
  "ISC",  # flake8-implicit-str-concat
  "LOG",  # flake8-logging
  "N",    # pep8-naming
  "PERF", # perflint
  "PIE",  # flake8-pie
  "PL",   # pylint
  "PT",   # flake8-pytest-style
  "PTH",  # flake8-use-pathlib
  "PYI",  # flake8-pyi
  "Q",    # flake8-quotes
  "RET",  # flake8-return
  "RSE",  # flake8-raise
  "RUF",  # ruff
  "S",    # flake8-bandit
  "SIM",  # flake8-simpify
  "SLF",  # flake8-self
  "SLOT", # flake8-slots
  "T100", # flake8-debugger
  "TRY",  # tryceratops
  "UP",   # pyupgrade
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]
lint.ignore = [
  "A005",   # allow to shadow stdlib and builtin module names
  "COM812", # trailing comma, conflicts with `ruff format`
  # Different doc rules that we don't really care about:
  "D100",
  "D104",
  "D106",
  "D203",
  "D212",
  "D401",
  "D404",
  "D405",
  "ISC001",  # implicit string concat conflicts with `ruff format`
  "ISC003",  # prefer explicit string concat over implicit concat
  "PLR09",   # we have our own complexity rules
  "PLR2004", # do not report magic numbers
  "PLR6301", # do not require classmethod / staticmethod when self not used
  "TRY003",  # long exception messages from `tryceratops`
  "RUF001",  # Я использую русский
]
lint.per-file-ignores."*.pyi" = [ "D103" ]
lint.per-file-ignores."tests/*.py" = [
  "RUF029", # allow async functions to not use `await`
  "S101",   # asserts
  "S105",   # hardcoded passwords
  "S404",   # subprocess calls are for tests
  "S603",   # do not require `shell=True`
  "S607",   # partial executable paths
]
lint.external = [ "WPS" ]
lint.flake8-quotes.inline-quotes = "single"
lint.mccabe.max-complexity = 6
lint.pep8-naming.staticmethod-decorators = [ "law_definition", "staticmethod" ]
lint.pydocstyle.convention = "google"
